Cache lists in Vuex store (#803)
* Cache lists in Vuex store

Instead of requesting list data from Amphora each time an
autocomplete behavior needs it, check to see if it is in the Vuex
store, and get it from the Vuex store if available.
If it's not available, dispatch an action which causes the lists
to be fetched.

Track pending/success/failure state of the lists request. Not
doing anything with it yet, but it's there if we want to reflect
this state in the UI or use Vue devtools to see what's happening.

Doesn't buy us much beyond cleanliness, since the browser will
probably cache the results of the Amphora calls anyway, but
it's a little cleaner to keep stuff in the store.

In this commit, I also added the first module to the Vuex store.
Seems like an easier way to divide up the Vuex stuff, and is
similar to `combineReducers` in Redux.

I deleted the existing `lists` stuff in the store so as not to
have name conflicts. As far as I can tell, that stuff wasn't used.

*  Tests for caching lists in Vuex store