ON-2506: allow for $el to be undefined
$el will be undefined when certain kiln inputs are used outside
the overlay.  Using kiln inputs outside the overlay is helpful
when writing a simple custom interface whose inputs should be
familiar to the editor.  Ideally the kiln inputs would be generic
enough to be used outside the context of kiln, however that is
surely a "nice to have" rather than a necessary feature.

Specifically the simple-list component caused this issue.