[wip] provides proof-of-concept for invalid link removal
This commit provides a working sub-class of Link that prevents the
creation of invalid anchor tags.

The validation is accomplished in the `SafeLink.optimize` method.
I was unable to get validation working in the `SafeLink.create` step
due to some actions Parchment takes after `create` to build the linked
list.

We need to determine how to organize custom Quill classes and what the
best practice is for loading them.

As an alternative, we could utilize the `SafeLink.create` method to set
an attribute on the anchor tag that marks the link as invalid (and
provide visual feedback via CSS? plus use that attribute as
validation?). I'll admit, I'm pretty out of my element when it comes to
front-end work so I would need to run that by the experts.