limits invalid-links check to just clay-(paragraph|subheader)
We don't need to run link validation on every component + every
property on the page. Our primary concern is links created with Quill
which we can target using generic Clay components.

The limit on this filter can be lifted if the use case presents itself,
but it's better to keep this number of affected components small to
start.