drops node10 test/build in CI
Since we are transpiling Kiln on release we probably don't need to
guarantee the ability to build with Node10 (which is EOL in a few
months). Removing that build compatability requirement will allow us to
use `string.prototype.matchAll` (needed for link validation).

The alternative would be to add another dependency or come up with our
own implementation in the short term.

All current major browsers support `.matchAll` so it's not worth the
hassle IMO.