fixes URL validation method for cross-browser support
In Firefox this will throw an Error:

```
new URL("https://www. example .com ")
```

In Chrome the same action will result in:

```
https://www.%20example%20.com
```

As a result the try/catch pattern used for validation rarely flagged
links on Chrome.

The approach in this commit uses a lenient regular expression to
validate URLs. A URL just needs a protocol, a domain, and optionally a
path. The domain and protocol have a limited character set.