styles invalid links in red, removes links on whitespace/punctuation
This commit addresses two UX issues with Kiln link validation.

**Invalid Link Styling**

Prior to this commit the only visual "hint" identifying invalid links
was the list of invalid components in the Kiln sidebar. By highlighting
links in red these errors will be more obvious to users.

:warning: I probably didn't organize this css change properly.

**Link Removal**

QuillJS requires the _full_ text of a link to be highlighted in order to
remove it. This can lead to little bits of punctuation or whitespace
retaining their hyperlink status. Identifying these single-character
links is especially frustrating if link validation is preventing a page
from being published.

This commit will automatically strip `<a>` tags that contain only a
single non-alphanumeric character as text.