Adds link validation for clay-paragraph, clay-subheader (#1505)
* [wip] provides proof-of-concept for invalid link removal

This commit provides a working sub-class of Link that prevents the
creation of invalid anchor tags.

The validation is accomplished in the `SafeLink.optimize` method.
I was unable to get validation working in the `SafeLink.create` step
due to some actions Parchment takes after `create` to build the linked
list.

We need to determine how to organize custom Quill classes and what the
best practice is for loading them.

As an alternative, we could utilize the `SafeLink.create` method to set
an attribute on the anchor tag that marks the link as invalid (and
provide visual feedback via CSS? plus use that attribute as
validation?). I'll admit, I'm pretty out of my element when it comes to
front-end work so I would need to run that by the experts.

* [wip] provides proof-of-concept for invalid link validator

* [wip] restores old sanitize behavior, removes about:blank

* fixes tests

The tests expected a '…' character in the preview text.
This commit adds that character.

* limits invalid-links check to just clay-(paragraph|subheader)

We don't need to run link validation on every component + every
property on the page. Our primary concern is links created with Quill
which we can target using generic Clay components.

The limit on this filter can be lifted if the use case presents itself,
but it's better to keep this number of affected components small to
start.

* drops node10 test/build in CI

Since we are transpiling Kiln on release we probably don't need to
guarantee the ability to build with Node10 (which is EOL in a few
months). Removing that build compatability requirement will allow us to
use `string.prototype.matchAll` (needed for link validation).

The alternative would be to add another dependency or come up with our
own implementation in the short term.

All current major browsers support `.matchAll` so it's not worth the
hassle IMO.

* uses valid-url package to validate URLs

* fixes URL validation method for cross-browser support

In Firefox this will throw an Error:

```
new URL("https://www. example .com ")
```

In Chrome the same action will result in:

```
https://www.%20example%20.com
```

As a result the try/catch pattern used for validation rarely flagged
links on Chrome.

The approach in this commit uses a lenient regular expression to
validate URLs. A URL just needs a protocol, a domain, and optionally a
path. The domain and protocol have a limited character set.

* fixes validation for anchors + mailto, allows same-site links

Validation that disallows same-site links might be too strict for
legacy content. Although protocol-less links cannot be created through
Kiln, Clay users can easily create `href="/section/page"`
links by importing content through the API.

This commit allows same-site links to avoid situations where old pages
cannot be re-published through the UI.

This commit does **not** allow same-site links that are not anchored
with a leading "/".

* uses querySelector to extract href values from text blobs

This commit replaces the dubious regex extraction of <a href=""> values
with a reliable method based on the document query selector.

Co-authored-by: jz <macgyver@users.noreply.github.com>

* [wip] attempts to use quill to identify invalid links

* styles invalid links in red, removes links on whitespace/punctuation

This commit addresses two UX issues with Kiln link validation.

**Invalid Link Styling**

Prior to this commit the only visual "hint" identifying invalid links
was the list of invalid components in the Kiln sidebar. By highlighting
links in red these errors will be more obvious to users.

:warning: I probably didn't organize this css change properly.

**Link Removal**

QuillJS requires the _full_ text of a link to be highlighted in order to
remove it. This can lead to little bits of punctuation or whitespace
retaining their hyperlink status. Identifying these single-character
links is especially frustrating if link validation is preventing a page
from being published.

This commit will automatically strip `<a>` tags that contain only a
single non-alphanumeric character as text.

* restores regex URL validation

Using `new URL` was too unpredictable across environments.
We were seeing tests fail in CircleCI, but pass locally (in `node`).

Paired with variability in browser behavior, I'm not sure how we can
adequately test a `new URL` implementation without a selenium test suite
(or something similar).

The regular expression method accepts:

* Anchor links
* Relative links to the site root
* Any link with `<protocol>:<path>`

More generally, the expression rejects anything containing spaces.

* fixes tests for invalid links

The invalid link tests were still expecting the link-parsing to be
performed during the validation function, but that responsibility was
moved to QuillJS. This commit adjusts the tests to look for the new
class.

:warning: This means we don't have any actual tests for link validation
in our custom link class.

* adds tests for isValid url validation method

* updates function docstring to reflect new behavior

* 10.1.0-0

* reverts pre-release versioning before merge

Co-authored-by: jz <macgyver@users.noreply.github.com>